Dumped testing project, but.
Set out to eliminate one sbt project. Did that and managed
to create two more. Sigh.
