+++ .ci/repl.txt
@@ -1,6 +1,6 @@
-val xs = 1 to 20 splitAt 10
+val xs = 1 to 20 splitAfter 10.size
 xs mapLeft (_ dropRight 8) join
 xs.zip filterRight (_ % 3 == 0)
-val ys = zipCross(1 to 3, vec("a", "bb", "ccc"))
-val same = ys filter (_ === _.length)
-same.rights joinWith '/'
+val ys = zipCross(1 to 3, view("a", "bb"))
+val zs = ys eqBy (x => x, _.length)
+zs.rights joinWith '/'
