+++ build.sbt
@@ -1,33 +1,42 @@
 import psp.PspStd._
 
-lazy val std = project.setup settings (
-  description :=  "psp's non-standard standard library"
-)
-
 lazy val root = (
-  project.root.setup aggregate std dependsOn (std, testing) settings (
-                       coursierVerbosity  :=  0,
-                     libraryDependencies  +=  ammonite,
-                      console in Compile <<=  ammoniteTask,
-   unmanagedSourceDirectories in Compile  +=  buildBase.value / "project" / "repl",
-                            watchSources <++= testing.allSources,
-                            watchSources <++= benchmark.allSources,
-                                    test <<=  test in testing
+  project.root.setup aggregate std dependsOn (std, benchmark) settings (
+      coursierVerbosity :=  0,
+     console in Compile <<= console in Compile in repl
   )
 )
 
-lazy val testing = project.noArtifacts.setup dependsOn std settings (
-                 testOptions +=  Tests.Argument(TestFrameworks.ScalaCheck, "-verbosity", "1"),
-                 testOptions +=  Tests.Argument(TestFrameworks.JUnit, "-s"),
-   parallelExecution in Test :=  false,
-                 logBuffered :=  false,
-                        test <<= test in Test,
-         libraryDependencies ++= testDependencies
-)
-lazy val benchmark = project.noArtifacts.setup dependsOn std enablePlugins JmhPlugin settings (
-  libraryDependencies ++= ( if (scalaBinaryVersion.value == "2.11") Seq(scoverageRuntime) else Seq() )
+lazy val std = project.setup settings (
+                description :=  "psp's non-standard standard library",
+                testOptions +=  Tests.Argument(TestFrameworks.ScalaCheck, "-verbosity", "1"),
+                testOptions +=  Tests.Argument(TestFrameworks.JUnit, "-s"),
+  parallelExecution in Test :=  false,
+        logBuffered in Test :=  false,
+        libraryDependencies ++= testDependencies map (_ % "test")
 )
 
+lazy val repl = project.noArtifacts.setup dependsOn std settings (
+     libraryDependencies += ammonite,
+   outputStrategy in run := Some(StdoutOutput),
+             fork in run := true,
+     connectInput in run := true,
+               mainClass := Some("psprepl.Main"),
+      console in Compile := (run in Compile).toTask(ammoniteArgs mkString (" ", " ", "")).value
+)
+
+lazy val macros = project.noArtifacts dependsOn (std % "compile->compile;test->test") settings (
+                 name :=  "psp-macros",
+  libraryDependencies +=  "org.scala-lang" % "scala-reflect" % scalaVersion.value,
+              version :=  "0.6.2-SNAPSHOT",
+         scalaVersion :=  "2.11.8",
+         organization :=  "org.improving",
+        scalacOptions ++= wordSeq("-language:experimental.macros -Yno-predef")
+)
+
+lazy val benchmark = project.noArtifacts.setup dependsOn std enablePlugins JmhPlugin settings (
+  libraryDependencies <++= (scalaBinaryVersion) { case "2.11" => Seq(scoverageRuntime) ; case _ => Nil }
+)
 
 addCommandAlias("bench-min", "benchmark/jmh:run -i 1 -wi 1 -f1 -t1")
 addCommandAlias("cover", "; clean ; coverage ; bench-min ; test ; coverageReport")
