+++ macros/src/main/scala/package.scala
@@ -1,0 +1,9 @@
+package psp
+package std
+
+import exp._
+
+package object macros {
+  def typechecked(expr: String): Typechecked                  = macro Typechecker.typechecked
+  def typecheckedLines(exprs: String): sciVector[Typechecked] = macro Typechecker.typecheckedLines
+}
