+++ macros/src/test/scala/generate.scala
