+++ project/plugins.sbt
@@ -1,7 +1,6 @@
   scalacOptions += "-deprecation"
 ivyLoggingLevel := UpdateLogging.Quiet
 
-addSbtPlugin("org.scoverage"       % "sbt-scoverage"           %   "1.3.5")
 addSbtPlugin("com.jsuereth"        % "sbt-pgp"                 %   "1.0.0")
 addSbtPlugin("org.xerial.sbt"      % "sbt-sonatype"            %    "1.1")
 addSbtPlugin("io.get-coursier"     % "sbt-coursier"            % "1.0.0-M12")
