+++ std/src/test/scala/main/Laws.scala
