+++ std/src/test/scala/tests/CollectionsSpec.scala
