+++ std/src/test/scala/tests/HeytingTests.scala
