+++ std/src/test/scala/tests/InferenceSpec.scala
