+++ std/src/test/scala/tests/MiscSpec.scala
