+++ std/src/test/scala/tests/TypeClassSpec.scala
