+++ std/src/test/scala/tests/ViewsVsEager.scala
