+++ std/src/test/scala/tests/XsViewTests.scala
